# docs/products/zhaoxi/openapi/app_v1.json
{
  "components": {
    "schemas": {
...
      "WorldEnvelope_HumanConversationListData_": {
        "properties": {
          "code": {
            "examples": [
              "ok"
            ],
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/HumanConversationListData"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "server_time": {
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "WorldEnvelope[HumanConversationListData]",
        "type": "object"
      },
      "WorldEnvelope_HumanReportOptionsData_": {
        "properties": {
          "code": {
            "examples": [
              "ok"
            ],
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/HumanReportOptionsData"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "server_time": {
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "WorldEnvelope[HumanReportOptionsData]",
        "type": "object"
      },
...
      "WorldErrorEnvelope": {
        "description": "世界类端点的失败信封：无 ``data``，``message`` 恒为 null，分支只看 ``code``。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "server_time": {
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id",
          "server_time"
        ],
        "title": "WorldErrorEnvelope",
        "type": "object"
      },
...
      "app__products__zhaoxi__api__companion_world_human_chat__EmptyPayload": {
        "additionalProperties": false,
        "description": "拒绝 participant/visit/world/account 注入字段。",
        "properties": {},
        "title": "EmptyPayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "朝夕相伴移动端 App 对接的 /v1 契约。公网入口 https://ai4company.top/api/v1/，服务端同时挂载对应公网路径；nginx 原 URI 透传。落地口径见 docs/products/zhaoxi/app_api_handoff.md。",
    "title": "朝夕相伴 App 客户端 API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/human-conversations": {
      "get": {
        "operationId": "list_human_conversations_v1_human_conversations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldEnvelope_HumanConversationListData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "未登录或 token 失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "账号或世界被停用"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "资源不存在或能力被关闭"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "状态冲突（只读、并发、容量）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "参数校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "触发限流"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "安全复核或异步任务暂不可用"
          }
        },
        "summary": "List Human Conversations",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/report-options": {
      "get": {
        "description": "举报原因受控表；随读门控开放，`human_chat_send=false` 时仍可举报。",
        "operationId": "list_human_report_options_v1_human_conversations_report_options_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldEnvelope_HumanReportOptionsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "未登录或 token 失效"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "账号或世界被停用"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "资源不存在或能力被关闭"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "状态冲突（只读、并发、容量）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "参数校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "触发限流"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldErrorEnvelope"
                }
              }
            },
            "description": "安全复核或异步任务暂不可用"
          }
        },
        "summary": "List Human Report Options",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/{conversation_id}/block": {
      "post": {
        "operationId": "block_human_counterpart_v1_human_conversations__conversation_id__block_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/app__products__zhaoxi__api__companion_world_human_chat__EmptyPayload"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Block Human Counterpart V1 Human Conversations  Conversation Id  Block Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block Human Counterpart",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/{conversation_id}/entry": {
      "delete": {
        "operationId": "hide_human_conversation_v1_human_conversations__conversation_id__entry_delete",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Hide Human Conversation V1 Human Conversations  Conversation Id  Entry Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hide Human Conversation",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/{conversation_id}/messages": {
      "get": {
        "operationId": "list_human_messages_v1_human_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Human Messages V1 Human Conversations  Conversation Id  Messages Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Human Messages",
        "tags": [
          "companion-world-human-chat"
        ]
      },
      "post": {
        "operationId": "send_human_message_v1_human_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendHumanMessagePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Send Human Message V1 Human Conversations  Conversation Id  Messages Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Human Message",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/{conversation_id}/read": {
      "post": {
        "operationId": "read_human_conversation_v1_human_conversations__conversation_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/app__products__zhaoxi__api__companion_world_human_chat__EmptyPayload"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Human Conversation V1 Human Conversations  Conversation Id  Read Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Human Conversation",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
    "/v1/human-conversations/{conversation_id}/report": {
      "post": {
        "operationId": "report_human_conversation_v1_human_conversations__conversation_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanReportPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Report Human Conversation V1 Human Conversations  Conversation Id  Report Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Human Conversation",
        "tags": [
          "companion-world-human-chat"
        ]
      }
    },
...
  }
}